backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
opencv-python==4.9.0.80
mediapipe==0.10.14
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.0
pandas==2.2.2
Pillow==10.3.0
httpx==0.27.0
